Refactor errors app and mimetype errors middleware
